apps/worker/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
numpy==1.26.4
trimesh==4.1.7
boto3==1.34.44
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.26.0
python-dotenv==1.0.1
scipy==1.12.0
meshio==5.3.5
ezdxf==1.1.4